Extracted a base class to the library project
