+++ peppol-smp-server-library/pom.xml
@@ -102,6 +102,10 @@
       <groupId>com.helger</groupId>
       <artifactId>ph-settings</artifactId>
     </dependency>
+    <dependency>
+      <groupId>com.helger</groupId>
+      <artifactId>ph-jaxb</artifactId>
+    </dependency>
 
     <dependency>
       <groupId>javax.servlet</groupId>
