+++ peppol-smp-server-webapp/src/main/java/com/helger/peppol/smpserver/rest/ServiceMetadataInterface.java
@@ -48,6 +48,7 @@ import com.helger.commons.xml.transform.XMLTransformerFactory;
 import com.helger.peppol.smp.ServiceMetadataType;
 import com.helger.peppol.smp.SignedServiceMetadataType;
 import com.helger.peppol.smpserver.SMPServerConfiguration;
+import com.helger.peppol.smpserver.jaxb.MarshallerSMPSignedServiceMetadataType;
 import com.helger.peppol.smpserver.restapi.SMPServerAPI;
 import com.helger.peppol.smpserver.security.SMPKeyManager;
 import com.helger.photon.core.app.CApplication;
@@ -86,7 +87,7 @@ public final class ServiceMetadataInterface
                                                                                                                                  sDocumentTypeID);
 
       // Convert to DOM document
-      final MarshallerSignedServiceMetadataType aMarshaller = new MarshallerSignedServiceMetadataType ();
+      final MarshallerSMPSignedServiceMetadataType aMarshaller = new MarshallerSMPSignedServiceMetadataType ();
       final Document aDoc = aMarshaller.write (ret);
 
       // Sign the document
